lab4/requirements.txt
annotated-types==0.6.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
paho-mqtt==1.6.1
pydantic==2.6.1
pydantic_core==2.16.2
requests==2.31.0
typing_extensions==4.9.0
urllib3==2.2.0
psycopg2-binary==2.9.3
